requirements.txt
Django==5.2.7
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
django-allauth>=0.51,<1.0
django-widget-tweaks==1.5.0
stripe==2.70.0
